Added gift request routes and model, added pages

+++ server/models/GiftRequest.js
@@ -1,4 +1,5 @@
 import mongoose from "mongoose";
+import {Decimal128} from "mongodb";
 
 const GiftRequestSchema = new mongoose.Schema({
     gift: {
@@ -11,7 +12,7 @@ const GiftRequestSchema = new mongoose.Schema({
     },
     category:           { type: String,         required: true },
     website:            { type: String,         required: true },
-    price:              { type: Float64Array,   required: true },
+    price:              { type: Decimal128,     required: true },
     status:             { type: String,         default: "purchased" },
     date_of_purchase:   { type: Date,           required: true },
     date_of_delivery:   { type: Date,           required: true },
